cleaning up the code and prepping it to connect to UI
I updated it with the code I had previously pushed to github (I forgot I had made the config.py file, so I brought it in and updated everything).

I also deleted the user_input and contact_link files because I know I won't be needed them.
